@@ -5,15 +5,15 @@
   "devDependencies": {
     "@types/jest": "^22.2.3",
     "@types/node": "^10.1.3",
-    "@types/react": "^16.3.14",
-    "@types/react-native": "^0.55.16",
+    "@types/react": "^16.8.23",
+    "@types/react-native": "^0.57.65",
     "@types/react-test-renderer": "^16.0.1",
-    "jest-expo": "^32.0.0",
+    "jest-expo": "^34.0.0",
     "react-native-scripts": "1.14.0",
     "react-native-typescript-transformer": "^1.2.11",
     "react-test-renderer": "^16.3.1",
     "ts-jest": "^22.4.6",
-    "typescript": "^3.3.1"
+    "typescript": "^3.4.5"
   },
   "main": "node_modules/expo/AppEntry.js",
   "scripts": {
@@ -48,21 +48,31 @@
   },
   "dependencies": {
     "@bible-engine/core": "file:../core",
+    "@expo/vector-icons": "^10.0.4",
+    "babel-preset-expo": "^5.2.0",
     "elasticlunr": "^0.9.5",
-    "expo": "^32.0.0",
-    "expo-analytics": "^1.0.8",
-    "react": "16.5.0",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-32.0.0.tar.gz",
+    "expo": "^34.0.3",
+    "expo-analytics": "^1.0.11",
+    "expo-asset": "^6.0.0",
+    "expo-constants": "^6.0.0",
+    "expo-file-system": "^6.0.2",
+    "expo-font": "^6.0.1",
+    "expo-keep-awake": "^6.0.0",
+    "expo-sqlite": "^6.0.0",
+    "react": "16.8.3",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-34.0.0.tar.gz",
     "react-native-collapsible": "^1.4.0",
     "react-native-console-time-polyfill": "^1.2.1",
+    "react-native-dotenv": "^0.2.0",
     "react-native-elements": "^1.1.0",
+    "react-native-gesture-handler": "^1.3.0",
     "react-native-indicators": "^0.13.0",
     "react-native-loading-placeholder": "^0.0.6",
     "react-native-paper": "^2.15.2",
     "react-native-simple-store": "^1.3.0",
     "react-native-snap-carousel": "^3.7.5",
     "react-native-sqlite-storage": "^3.3.5",
-    "sentry-expo": "^1.12.0",
+    "sentry-expo": "~1.13.0",
     "tslib": "^1.9.1"
   }
 }
