@@ -66,7 +66,7 @@ export default class CrossReference extends React.PureComponent<Props, State> {
   }
 
   async getVerseContents(refs: IBibleCrossReference[]) {
-    const verseContents = refs.map(ref => ref.range).map(range => '');
+    const verseContents = await this.props.database.getVerseContents(refs);
     if (this.mounted) {
       this.setState({
         ...this.state,
@@ -122,7 +122,7 @@ export default class CrossReference extends React.PureComponent<Props, State> {
           isVisible={this.state.popoverIsVisible}
           fromView={this.touchable}
           popoverStyle={styles.popover__background_container}
-          onClose={() => this.closePopover()}
+          onRequestClose={() => this.closePopover()}
         >
           {this.renderPopoverContent()}
         </Popover>
