@@ -31,7 +31,7 @@ interface State {
   popoverIsVisible: boolean;
 }
 
-export default class CrossReference extends React.PureComponent<Props, State> {
+export default class Footnote extends React.PureComponent<Props, State> {
   touchable: any;
   state = {
     popoverIsVisible: false
@@ -88,7 +88,7 @@ export default class CrossReference extends React.PureComponent<Props, State> {
           isVisible={this.state.popoverIsVisible}
           fromView={this.touchable}
           popoverStyle={styles.popover__background_container}
-          onClose={() => this.closePopover()}
+          onRequestClose={() => this.closePopover()}
         >
           {this.renderPopoverContent()}
         </Popover>
